+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add `LiteSVM::with_precompiles`.
+
 ## [0.3.0] - 2024-10-12
 
 ### Added